Move dummy api to own module
